+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## v0.5.13 [unrelease]
+## v0.6.1 [unreleased]
+
+### Bugfixes
+
+## v0.6.0 [2014-05-02]
 
 ### Feature
 
@@ -10,6 +14,7 @@
 - [Issue #469](https://github.com/influxdb/influxdb/issues/469). Drop continuous queries when a database is dropped
 - [Issue #431](https://github.com/influxdb/influxdb/issues/431). Don't log to standard output if a log file is specified in the config file
 - [Issue #483](https://github.com/influxdb/influxdb/issues/483). Return 409 if a database already exist (Thanks, Edward Muller)
+- [Issue #486](https://github.com/influxdb/influxdb/issues/486). Columns used in the target of continuous query shouldn't be inserted in the time series
 - [Issue #490](https://github.com/influxdb/influxdb/issues/490). Database user password's cannot be changed (Thanks, Edward Muller)
 - [Issue #495](https://github.com/influxdb/influxdb/issues/495). Enforce write permissions properly